Replace single PlayerCharacterItems per character with a PlayerCharacterItem per character item
- Filtering on JSONB columns and extracting things from arrays is very slow, anything requiring database filtering will need to be an actual table
