+++ src/Wowthing.Lib/Models/Player/PlayerCharacterItem.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using Microsoft.EntityFrameworkCore;
+using Wowthing.Lib.Enums;
+
+namespace Wowthing.Lib.Models.Player
+{
+    [Index(nameof(CharacterId), nameof(ItemId), nameof(Location))]
+    public class PlayerCharacterItem
+    {
+        // Fields are ordered from largest to smallest for database table size reasons. Postgres doesn't go
+        // smaller than a short (2 bytes), sadly.
+        [Key]
+        public long Id { get; set; }
+        
+        [ForeignKey("Character")]
+        public int CharacterId { get; set; }
+        public PlayerCharacter Character { get; set; }
+        
+        public int ItemId { get; set; }
+        public int Count { get; set; }
+
+        public ItemLocation Location { get; set; }
+        public short BagId { get; set; }
+        public short Slot { get; set; }
+
+        public short Context { get; set; }
+        public short EnchantId { get; set; }
+        public short ItemLevel { get; set; }
+        public short Quality { get; set; }
+        public short SuffixId { get; set; }
+
+        public List<int> Gems { get; set; }
+        public List<short> BonusIds { get; set; }
+    }
+}
